requirements.txt
Django==1.11.4
django-bootstrap==0.2.4
django-bootstrap3==9.0.0
django-captcha==1.0.0
django-ckeditor==5.3.0
django-cleanup==1.0.1
django-crispy-forms==1.6.1
django-el-pagination==3.0.1
django-js-asset==0.1.1
django-markdown==0.8.4
django-markdown-deux==1.0.5
django-pagedown==0.1.3
django-patterns==0.0.3
django-recaptcha==1.3.1
django-simple-blog==0.2.1
django-simple-math-captcha==1.0.7
gevent==1.2.2
greenlet==0.4.12
Markdown==2.6.9
markdown2==2.3.4
olefile==0.44
Pillow==4.2.1
psycopg2==2.6.2
pytz==2017.2
